% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{json_array}
\alias{json_array}
\title{Create a schema for an array/vector}
\usage{
json_array(items, .optional = FALSE)
}
\arguments{
\item{items}{Schema definition for array elements}

\item{.optional}{Logical; if TRUE, this array can be null}
}
\value{
A schema definition object
}
\description{
Create a schema for an array/vector
}
\examples{
# Array of integers
json_array(json_integer())

# Array of user objects
json_array(json_object(
  name = json_string(),
  age = json_integer()
))
}
