% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{json_boolean}
\alias{json_boolean}
\title{Create a schema for a logical/boolean value}
\usage{
json_boolean(.optional = FALSE, .default = NULL)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null and a missing field
will not cause an error}

\item{.default}{Default value to use when field is missing. Only applies to
required fields (.optional = FALSE). Must be a single logical value (TRUE or FALSE)}
}
\value{
A schema definition object
}
\description{
Create a schema for a logical/boolean value
}
\examples{
json_boolean()
json_boolean(.optional = TRUE)
json_boolean(.optional = TRUE, .default = FALSE)
}
