% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.BuiltSchema}
\alias{print.BuiltSchema}
\title{Print method for built schema objects}
\usage{
\method{print}{BuiltSchema}(x, ...)
}
\arguments{
\item{x}{A built schema object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for built schema objects
}
