% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.llmjson_schema}
\alias{print.llmjson_schema}
\title{Print method for schema objects}
\usage{
\method{print}{llmjson_schema}(x, ..., indent = 0)
}
\arguments{
\item{x}{A schema object}

\item{...}{Additional arguments (ignored)}

\item{indent}{Current indentation level}
}
\description{
Print method for schema objects
}
