% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_any}
\alias{s_any}
\title{Create a schema that allows any type}
\usage{
s_any(.optional = FALSE)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null}
}
\value{
A schema definition object
}
\description{
Create a schema that allows any type
}
\examples{
s_any()
}
