% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_array}
\alias{s_array}
\title{Create a schema for an array/vector}
\usage{
s_array(items, .optional = FALSE)
}
\arguments{
\item{items}{Schema definition for array elements}

\item{.optional}{Logical; if TRUE, this array can be null}
}
\value{
A schema definition object
}
\description{
Create a schema for an array/vector
}
\examples{
# Array of integers
s_array(s_integer())

# Array of user objects
s_array(s_map(
  name = s_string(),
  age = s_integer()
))
}
