% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_date}
\alias{s_date}
\title{Create a schema for a Date value}
\usage{
s_date(.optional = FALSE, .default = NULL, .format = "iso8601")
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null and a missing field
will not cause an error}

\item{.default}{Default value to use when field is missing and .optional = TRUE.
Must be a single Date object (e.g., Sys.Date(), as.Date("2024-01-01"))}

\item{.format}{Character vector specifying date format(s). Can be:
\itemize{
\item A named format: "iso8601" (\\%Y-\\%m-\\%d), "us_date" (\\%m/\\%d/\\%Y),
"eu_date" (\\%d/\\%m/\\%Y)
\item A custom strptime format string (e.g., "\\%d-\\%m-\\%Y")
\item A vector of formats to try in order
}
Defaults to "iso8601" (\\%Y-\\%m-\\%d)}
}
\value{
A schema definition object
}
\description{
This schema parses date strings from JSON into R Date objects. It supports
both named format presets and custom strptime format strings.
}
\examples{
s_date()
s_date(.optional = TRUE)
s_date(.format = "us_date")
s_date(.format = "\%d-\%m-\%Y")
s_date(.format = c("iso8601", "us_date"))
}
