% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_double}
\alias{s_double}
\title{Create a schema for a double/numeric value}
\usage{
s_double(.optional = FALSE, .default = NULL)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null}

\item{.default}{Default value to use if field is missing. If provided,
the field is not required}
}
\value{
A schema definition object
}
\description{
Create a schema for a double/numeric value
}
\examples{
s_double()
s_double(.optional = TRUE)
s_double(.default = 0.0)
}
