% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_integer}
\alias{s_integer}
\title{Create a schema for an integer value}
\usage{
s_integer(.optional = FALSE, .default = NULL)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null}

\item{.default}{Default value to use if field is missing. If provided,
the field is not required}
}
\value{
A schema definition object
}
\description{
Create a schema for an integer value
}
\examples{
s_integer()
s_integer(.optional = TRUE)
s_integer(.default = 0L)
}
