% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_logical}
\alias{s_logical}
\title{Create a schema for a logical/boolean value}
\usage{
s_logical(.optional = FALSE, .default = NULL)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null and a missing field
will not cause an error}

\item{.default}{Default value to use when field is missing and .optional = TRUE.
Must be a single logical value (TRUE or FALSE)}
}
\value{
A schema definition object
}
\description{
Create a schema for a logical/boolean value
}
\examples{
s_logical()
s_logical(.optional = TRUE)
s_logical(.optional = TRUE, .default = FALSE)
}
