% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_map}
\alias{s_map}
\title{Create a schema for a JSON object/map}
\usage{
s_map(..., .optional = FALSE)
}
\arguments{
\item{...}{Named arguments defining the schema for each field.
Each argument should be a schema builder (s_integer, s_double, s_string, etc.)}

\item{.optional}{Logical; if TRUE, this object can be null}
}
\value{
A schema definition object
}
\description{
Create a schema for a JSON object/map
}
\examples{
# Define a schema for a user object
schema <- s_map(
  name = s_string(),
  age = s_integer(),
  email = s_string()
)
}
