% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_posixct}
\alias{s_posixct}
\title{Create a schema for a POSIXct datetime value}
\usage{
s_posixct(.optional = FALSE, .default = NULL, .format = "iso8601", .tz = "UTC")
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null and a missing field
will not cause an error}

\item{.default}{Default value to use when field is missing and .optional = TRUE.
Must be a single POSIXct object (e.g., Sys.time(), as.POSIXct("2024-01-01 12:00:00"))}

\item{.format}{Character vector specifying datetime format(s). Can be:
\itemize{
\item A named format: "iso8601" (\\%Y-\\%m-\\%dT\\%H:\\%M:\\%S),
"iso8601z" (\\%Y-\\%m-\\%dT\\%H:\\%M:\\%SZ), "rfc822" (\\%a, \\%d \\%b \\%Y \\%H:\\%M:\\%S),
"us_datetime" (\\%m/\\%d/\\%Y \\%H:\\%M:\\%S), "eu_datetime" (\\%d/\\%m/\\%Y \\%H:\\%M:\\%S)
\item "unix" or "epoch" to parse numeric Unix timestamps (seconds since 1970-01-01)
\item "unix_ms" to parse millisecond timestamps
\item A custom strptime format string
\item A vector of formats to try in order
}
Defaults to "iso8601" (\\%Y-\\%m-\\%dT\\%H:\\%M:\\%S)}

\item{.tz}{Timezone to use for parsing. Defaults to "UTC"}
}
\value{
A schema definition object
}
\description{
This schema parses datetime strings from JSON into R POSIXct objects. It supports
both named format presets and custom strptime format strings, as well as numeric
Unix timestamps.
}
\examples{
s_posixct()
s_posixct(.optional = TRUE)
s_posixct(.format = "iso8601z")
s_posixct(.format = "unix")
s_posixct(.format = c("iso8601", "iso8601z"))
s_posixct(.tz = "America/New_York")
}
