% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{s_string}
\alias{s_string}
\title{Create a schema for a string value}
\usage{
s_string(.optional = FALSE, .default = NULL)
}
\arguments{
\item{.optional}{Logical; if TRUE, this value can be null}

\item{.default}{Default value to use if field is missing. If provided,
the field is not required}
}
\value{
A schema definition object
}
\description{
Create a schema for a string value
}
\examples{
s_string()
s_string(.optional = TRUE)
s_string(.default = "")
}
